+++ asyncz/schedulers/defaults.py
@@ -1,15 +1,14 @@
-from typing import Dict
-
-triggers: Dict[str, str] = {
+triggers: dict[str, str] = {
     "date": "asyncz.triggers.date:DateTrigger",
     "interval": "asyncz.triggers.interval:IntervalTrigger",
     "cron": "asyncz.triggers.cron.trigger:CronTrigger",
     "and": "asyncz.triggers.date:DateTrigger",
     "or": "asyncz.triggers.combining:OrTrigger",
+    "shutdown": "asyncz.triggers.shutdown:ShutdownTrigger",
 }
 
 
-executors: Dict[str, str] = {
+executors: dict[str, str] = {
     "debug": "asyncz.executors.debug:DebugExecutor",
     "pool": "asyncz.executors.pool:ThreadPoolExecutor",
     "threadpool": "asyncz.executors.pool:ThreadPoolExecutor",
@@ -17,7 +16,7 @@ executors: Dict[str, str] = {
     "asyncio": "asyncz.executors.asyncio:AsyncIOExecutor",
 }
 
-stores: Dict[str, str] = {
+stores: dict[str, str] = {
     "memory": "asyncz.stores.memory:MemoryTaskStore",
     "mongodb": "asyncz.stores.mongo:MongoDBStore",
     "redis": "asyncz.stores.redis:RedisStore",