+++ asyncz/triggers/__init__.py
@@ -3,6 +3,7 @@ from .combination import AndTrigger, OrTrigger
 from .cron import CronTrigger
 from .date import DateTrigger
 from .interval import IntervalTrigger
+from .shutdown import ShutdownTrigger
 
 __all__ = [
     "BaseCombinationTrigger",
@@ -12,4 +13,5 @@ __all__ = [
     "CronTrigger",
     "DateTrigger",
     "IntervalTrigger",
+    "ShutdownTrigger",
 ]